:sparkles: intergrityDeploy script added
