+++ app/code/community/Uecommerce/Mundipagg/controllers/StandardController.php
@@ -502,8 +502,8 @@ class Uecommerce_Mundipagg_StandardController extends Mage_Core_Controller_Front
         }
 
         //gathering environment and module info
-        $modmanFilePath = './app/code/community/Uecommerce/Mundipagg/etc/modman';
-        $integrityCheckFile = './app/code/community/Uecommerce/Mundipagg/etc/integrityCheck';
+        $modmanFilePath = './app/code/community/Uecommerce/Mundipagg/etc/integrity/modman';
+        $integrityCheckFile = './app/code/community/Uecommerce/Mundipagg/etc/integrity/integrityCheck';
 
         $installType = 'package';
 
@@ -522,7 +522,10 @@ class Uecommerce_Mundipagg_StandardController extends Mage_Core_Controller_Front
 
         //integrity check
         $integrityData = json_decode(file_get_contents($integrityCheckFile),true);
-        $files = $this->generateModuleFilesMD5s($modmanFilePath);
+
+        require_once './app/code/community/Uecommerce/Mundipagg/etc/integrity/IntegrityEngine.php';
+        $integrityEngine = new IntegrityEngine();
+        $files = $integrityEngine->generateModuleFilesMD5s($modmanFilePath);
 
         $newFiles = [];
         $unreadableFiles = [];
@@ -531,8 +534,7 @@ class Uecommerce_Mundipagg_StandardController extends Mage_Core_Controller_Front
 
         //validating files
         foreach ($files as $fileName => $md5) {
-            if (
-                $fileName === './app/code/community/Uecommerce/Mundipagg/etc/integrityCheck') {
+            if (substr($fileName, -strlen('integrityCheck')) == 'integrityCheck') {
                 //skip validation of integrityCheck file
                 continue;
             }
@@ -589,105 +591,5 @@ class Uecommerce_Mundipagg_StandardController extends Mage_Core_Controller_Front
         echo '<h3>phpnfo()</h3>';
         phpinfo();
 
-    }
-
-    protected function dirCheckSum($dir)
-    {
-        $files = scandir($dir);
-        $md5 = [];
-        foreach ($files as $file) {
-            if ($file !== '.' && $file !== '..') {
-                $file = $dir . DIRECTORY_SEPARATOR . $file;
-
-                $checkMethod = 'fileCheckSum';
-                if (is_dir($file)) {
-                    $checkMethod = 'dirCheckSum';
-                }
-
-                $md5[$file] = $this->$checkMethod($file);
-            }
-        }
-        return $md5;
-    }
-
-    public function generateModuleFilesMD5s($modmanFilePath) {
-
-        $modmanRawData = file_get_contents($modmanFilePath);
-
-        $rawLines = explode("\n",$modmanRawData);
-        $lines = [];
-        foreach ($rawLines as $rawLine) {
-            if (
-                substr($rawLine,0,1) === '#' ||
-                strlen($rawLine) === 0
-            ) {
-                continue;
-            }
-            $lines[] = array_values(array_filter(explode(' ',$rawLine)));
-        }
-        foreach ($lines as $index => $line) {
-            $elementName = './' . $line[1];
-            $checkMethod = 'fileCheckSum';
-            if (is_dir($elementName)) {
-                $checkMethod = 'dirCheckSum';
-            }
-            $lines[$index][] = $this->filterFileCheckSum($this->$checkMethod($elementName));
-        }
-        $files = [];
-        foreach($lines as $line) {
-            $files = array_merge($files,end($line));
-        }
-
-        //removing modman base files from generated hashs.
-        foreach($files as $filePath => $md5) {
-            if (strpos($filePath,'./.modman/') !== false) {
-                unset($files[$filePath]);
-            }
-        }
-
-        return $files;
-    }
-
-    protected function filterFileCheckSum($checkSumArray)
-    {
-        if(count($checkSumArray) === 1) {
-            return $checkSumArray;
-        }
-        $data = serialize($checkSumArray);
-        $data = explode('";s:32:"',$data);
-        $currentFile = null;
-        $currentMd5 = null;
-        $files = [];
-        foreach ($data as $line) {
-            $raw = explode('"',$line);
-            if( $currentFile ) {
-                $files[$currentFile] = $raw[0];
-                $currentFile = end($raw);
-                continue;
-            }
-            $currentFile = end($raw);
-        }
-        return $files;
-    }
-
-    protected function filterFileMd5($pathArray)
-    {
-        $files = [];
-        foreach ($pathArray as $path => $md5) {
-            if (!is_dir($path)) {
-                $files[$path] = $md5;
-            }
-            else {
-                $files = array_merge($files,$this->filterFileMd5($path));
-            }
-        }
-        return $files;
-    }
-
-    protected function fileCheckSum($file)
-    {
-        return  [
-            $file => md5_file($file)
-        ];
     }
 }
